+++ CHANGELOG.md
@@ -1,10 +1,30 @@
 # Changelog
 
+## [v3.2.0] - minor updates - 2019-03-25
+
+### Added
+
+- reverse() for ColorScheme
+- docs update
+
+### Changed
+
+-
+
+### Removed
+
+-
+
+### Deprecated
+
+-
+
 ## [v3.1.0] - minor changes - 2019-02-15
 
 ### Added
 
 - twilight added
+- reverse() for ColorScheme
 
 ### Changed